+++ tests/test_api.py
@@ -160,7 +160,7 @@ class TestMintNewDOI(object):
         mock_crud.read_package = MagicMock(return_value=None)
 
         api = DataciteClient()
-        doi = constants.XML_DICT['identifiers'][0]['identifier']
+        doi = constants.XML_DICT['doi']
         pkg_id = MagicMock()
 
         with pytest.raises(DataCiteError):
@@ -174,7 +174,7 @@ class TestMintNewDOI(object):
         mock_crud.read_package = MagicMock(return_value=None)
 
         api = DataciteClient()
-        doi = constants.XML_DICT['identifiers'][0]['identifier']
+        doi = constants.XML_DICT['doi']
         pkg_id = MagicMock()
 
         api.set_metadata(doi, constants.XML_DICT)
@@ -189,7 +189,7 @@ class TestMintNewDOI(object):
         mock_crud.read_package = MagicMock(return_value=None)
 
         api = DataciteClient()
-        doi = constants.XML_DICT['identifiers'][0]['identifier']
+        doi = constants.XML_DICT['doi']
         pkg_id = MagicMock()
 
         api.set_metadata(doi, constants.XML_DICT)
@@ -204,7 +204,7 @@ class TestMintNewDOI(object):
         mock_crud.read_package = MagicMock(return_value=MagicMock())
 
         api = DataciteClient()
-        doi = constants.XML_DICT['identifiers'][0]['identifier']
+        doi = constants.XML_DICT['doi']
         pkg_id = MagicMock()
 
         api.set_metadata(doi, constants.XML_DICT)
@@ -219,7 +219,7 @@ class TestMintNewDOI(object):
         mock_crud.read_package = MagicMock(return_value=MagicMock())
 
         api = DataciteClient()
-        doi = constants.XML_DICT['identifiers'][0]['identifier']
+        doi = constants.XML_DICT['doi']
         pkg_id = MagicMock()
 
         api.set_metadata(doi, constants.XML_DICT)